Flesh out tests for saga resulting messages behavior
